+++ tests/utils_tests.py
@@ -48,7 +48,6 @@ from superset.utils.core import (
     get_iterable,
     get_email_address_list,
     get_or_create_db,
-    get_since_until,
     get_stacktrace,
     json_int_dttm_ser,
     json_iso_dttm_ser,
@@ -57,15 +56,12 @@ from superset.utils.core import (
     merge_extra_filters,
     merge_request_params,
     parse_ssl_cert,
-    parse_human_timedelta,
     parse_js_uri_path_item,
-    parse_past_timedelta,
     split,
     TimeRangeEndpoint,
     validate_json,
     zlib_compress,
     zlib_decompress,
-    datetime_eval,
 )
 from superset.utils import schema
 from superset.views.utils import (
@@ -76,31 +72,6 @@ from superset.views.utils import (
 from tests.base_tests import SupersetTestCase
 
 from .fixtures.certificates import ssl_certificate
-
-
-def mock_parse_human_datetime(s):
-    if s == "now":
-        return datetime(2016, 11, 7, 9, 30, 10)
-    elif s == "today":
-        return datetime(2016, 11, 7)
-    elif s == "yesterday":
-        return datetime(2016, 11, 6)
-    elif s == "tomorrow":
-        return datetime(2016, 11, 8)
-    elif s == "Last year":
-        return datetime(2015, 11, 7)
-    elif s == "Last week":
-        return datetime(2015, 10, 31)
-    elif s == "Last 5 months":
-        return datetime(2016, 6, 7)
-    elif s == "Next 5 months":
-        return datetime(2017, 4, 7)
-    elif s in ["5 days", "5 days ago"]:
-        return datetime(2016, 11, 2)
-    elif s == "2018-01-01T00:00:00":
-        return datetime(2018, 1, 1)
-    elif s == "2018-12-31T23:59:59":
-        return datetime(2018, 12, 31, 23, 59, 59)
 
 
 def mock_to_adhoc(filt, expressionType="SIMPLE", clause="where"):
@@ -148,36 +119,6 @@ class TestUtils(SupersetTestCase):
         assert isinstance(base_json_conv(Decimal("1.0")), float) is True
         assert isinstance(base_json_conv(uuid.uuid4()), str) is True
         assert isinstance(base_json_conv(timedelta(0)), str) is True
-
-    @patch("superset.utils.core.datetime")
-    def test_parse_human_timedelta(self, mock_datetime):
-        mock_datetime.now.return_value = datetime(2019, 4, 1)
-        mock_datetime.side_effect = lambda *args, **kw: datetime(*args, **kw)
-        self.assertEqual(parse_human_timedelta("now"), timedelta(0))
-        self.assertEqual(parse_human_timedelta("1 year"), timedelta(366))
-        self.assertEqual(parse_human_timedelta("-1 year"), timedelta(-365))
-        self.assertEqual(parse_human_timedelta(None), timedelta(0))
-        self.assertEqual(
-            parse_human_timedelta("1 month", datetime(2019, 4, 1)), timedelta(30),
-        )
-        self.assertEqual(
-            parse_human_timedelta("1 month", datetime(2019, 5, 1)), timedelta(31),
-        )
-        self.assertEqual(
-            parse_human_timedelta("1 month", datetime(2019, 2, 1)), timedelta(28),
-        )
-        self.assertEqual(
-            parse_human_timedelta("-1 month", datetime(2019, 2, 1)), timedelta(-31),
-        )
-
-    @patch("superset.utils.core.datetime")
-    def test_parse_past_timedelta(self, mock_datetime):
-        mock_datetime.now.return_value = datetime(2019, 4, 1)
-        mock_datetime.side_effect = lambda *args, **kw: datetime(*args, **kw)
-        self.assertEqual(parse_past_timedelta("1 year"), timedelta(365))
-        self.assertEqual(parse_past_timedelta("-1 year"), timedelta(365))
-        self.assertEqual(parse_past_timedelta("52 weeks"), timedelta(364))
-        self.assertEqual(parse_past_timedelta("1 month"), timedelta(31))
 
     def test_zlib_compression(self):
         json_str = '{"test": 1}'
@@ -700,186 +641,6 @@ class TestUtils(SupersetTestCase):
         result8 = instance.test_method(1, 2, 3)
         self.assertEqual(instance.watcher, 4)
         self.assertEqual(result1, result8)
-
-    @patch("superset.utils.core.parse_human_datetime", mock_parse_human_datetime)
-    def test_get_since_until(self):
-        result = get_since_until()
-        expected = None, datetime(2016, 11, 7)
-        self.assertEqual(result, expected)
-
-        result = get_since_until(" : now")
-        expected = None, datetime(2016, 11, 7, 9, 30, 10)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("yesterday : tomorrow")
-        expected = datetime(2016, 11, 6), datetime(2016, 11, 8)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("2018-01-01T00:00:00 : 2018-12-31T23:59:59")
-        expected = datetime(2018, 1, 1), datetime(2018, 12, 31, 23, 59, 59)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("Last year")
-        expected = datetime(2015, 11, 7), datetime(2016, 11, 7)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("Last quarter")
-        expected = datetime(2016, 8, 7), datetime(2016, 11, 7)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("Last 5 months")
-        expected = datetime(2016, 6, 7), datetime(2016, 11, 7)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("Next 5 months")
-        expected = datetime(2016, 11, 7), datetime(2017, 4, 7)
-        self.assertEqual(result, expected)
-
-        result = get_since_until(since="5 days")
-        expected = datetime(2016, 11, 2), datetime(2016, 11, 7)
-        self.assertEqual(result, expected)
-
-        result = get_since_until(since="5 days ago", until="tomorrow")
-        expected = datetime(2016, 11, 2), datetime(2016, 11, 8)
-        self.assertEqual(result, expected)
-
-        result = get_since_until(time_range="yesterday : tomorrow", time_shift="1 day")
-        expected = datetime(2016, 11, 5), datetime(2016, 11, 7)
-        self.assertEqual(result, expected)
-
-        result = get_since_until(time_range="5 days : now")
-        expected = datetime(2016, 11, 2), datetime(2016, 11, 7, 9, 30, 10)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("Last week", relative_end="now")
-        expected = datetime(2016, 10, 31), datetime(2016, 11, 7, 9, 30, 10)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("Last week", relative_start="now")
-        expected = datetime(2016, 10, 31, 9, 30, 10), datetime(2016, 11, 7)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("Last week", relative_start="now", relative_end="now")
-        expected = datetime(2016, 10, 31, 9, 30, 10), datetime(2016, 11, 7, 9, 30, 10)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("previous calendar week")
-        expected = datetime(2016, 10, 31, 0, 0, 0), datetime(2016, 11, 7, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("previous calendar month")
-        expected = datetime(2016, 10, 1, 0, 0, 0), datetime(2016, 11, 1, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = get_since_until("previous calendar year")
-        expected = datetime(2015, 1, 1, 0, 0, 0), datetime(2016, 1, 1, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        with self.assertRaises(ValueError):
-            get_since_until(time_range="tomorrow : yesterday")
-
-    @patch("superset.utils.core.parse_human_datetime", mock_parse_human_datetime)
-    def test_datetime_eval(self):
-        result = datetime_eval("datetime('now')")
-        expected = datetime(2016, 11, 7, 9, 30, 10)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("datetime('today'  )")
-        expected = datetime(2016, 11, 7)
-        self.assertEqual(result, expected)
-
-        # Parse compact arguments spelling
-        result = datetime_eval("dateadd(datetime('today'),1,year,)")
-        expected = datetime(2017, 11, 7)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("dateadd(datetime('today'), -2, year)")
-        expected = datetime(2014, 11, 7)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("dateadd(datetime('today'), 2, quarter)")
-        expected = datetime(2017, 5, 7)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("dateadd(datetime('today'), 3, month)")
-        expected = datetime(2017, 2, 7)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("dateadd(datetime('today'), -3, week)")
-        expected = datetime(2016, 10, 17)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("dateadd(datetime('today'), 3, day)")
-        expected = datetime(2016, 11, 10)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("dateadd(datetime('now'), 3, hour)")
-        expected = datetime(2016, 11, 7, 12, 30, 10)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("dateadd(datetime('now'), 40, minute)")
-        expected = datetime(2016, 11, 7, 10, 10, 10)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("dateadd(datetime('now'), -11, second)")
-        expected = datetime(2016, 11, 7, 9, 29, 59)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("datetrunc(datetime('now'), year)")
-        expected = datetime(2016, 1, 1, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("datetrunc(datetime('now'), month)")
-        expected = datetime(2016, 11, 1, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("datetrunc(datetime('now'), day)")
-        expected = datetime(2016, 11, 7, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("datetrunc(datetime('now'), week)")
-        expected = datetime(2016, 11, 7, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("datetrunc(datetime('now'), hour)")
-        expected = datetime(2016, 11, 7, 9, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("datetrunc(datetime('now'), minute)")
-        expected = datetime(2016, 11, 7, 9, 30, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("datetrunc(datetime('now'), second)")
-        expected = datetime(2016, 11, 7, 9, 30, 10)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("lastday(datetime('now'), year)")
-        expected = datetime(2016, 12, 31, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("lastday(datetime('today'), month)")
-        expected = datetime(2016, 11, 30, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("holiday('Christmas')")
-        expected = datetime(2016, 12, 25, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval("holiday('Labor day', datetime('2018-01-01T00:00:00'))")
-        expected = datetime(2018, 9, 3, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval(
-            "holiday('Boxing day', datetime('2018-01-01T00:00:00'), 'UK')"
-        )
-        expected = datetime(2018, 12, 26, 0, 0, 0)
-        self.assertEqual(result, expected)
-
-        result = datetime_eval(
-            "lastday(dateadd(datetime('2018-01-01T00:00:00'), 1, month), month)"
-        )
-        expected = datetime(2018, 2, 28, 0, 0, 0)
-        self.assertEqual(result, expected)
 
     @patch("superset.utils.core.to_adhoc", mock_to_adhoc)
     def test_convert_legacy_filters_into_adhoc_where(self):
